Also move compact_level0
